pyproject.toml
[project]
name = "security-for-cve-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "duckduckgo-search>=8.1.1",
    "mcp[cli]==1.18.0",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
]
